Moving All handling of broker service from EA.D into the implementation of IVisualDiagnosticsLanguageService
