+++ src/Tools/ExternalAccess/VisualDiagnostics/Contracts/IVisualDiagnosticsServiceBroker.cs
@@ -9,13 +9,10 @@ using Microsoft.VisualStudio.Debugger.Contracts.HotReload;
 namespace Microsoft.CodeAnalysis.ExternalAccess.VisualDiagnostics.Contracts
 {
     /// <summary>
-    /// Facade interface for getting various service brokers
+    /// Facade interface for getting broker service successfully through MEF
     /// </summary>
     internal interface IVisualDiagnosticsBrokeredDebuggerServices
     {
         Task<IServiceBroker> GetServiceBrokerAsync();
-        Task<IHotReloadSessionNotificationService?> GetHotReloadSessionNotificationServiceAsync();
-        Task<IManagedHotReloadAgentManagerService?> GetManagedHotReloadAgentManagerServiceAsync();
-        Task<IManagedHotReloadService?> GetManagedHotReloadServiceAsync();
     }
 }
